Respect Previous BuildPlans
Previously, when the buildpack contributed entries to the BuildPlan, it would
write new Metadata regardless of what previous buildpacks contributed.  This
was problematic because if another buildpack contributed the same dependency
it would remove any configuration from the previous buildpack.  This change
updates the contribution so that it reads the previous dependency and only
adds/modifies metadata entries that are pertinent to it.  Any Metadata that
previously existed should be carried forward.

Signed-off-by: Ben Hale <bhale@pivotal.io>