Treat Empty String as Unset
Previously, the implementation of metadata made the decision to override the
values in riff.toml solely on whether a property had been set.  Due to
restrictions in the implementation of CRDs, it is not possible to determine
whether a value is unset or simply empty.

This change updates the implementation to treat an empty string value as
unset, even as it loses fidelity over the actual value.  This will prevent any
user from overriding the value in a riff.toml with an empty string in an
effort to get different behavior.

Signed-off-by: Ben Hale <bhale@vmware.com>